Move kindchecker to separate file
