+++ shared/src/main/scala/effekt/Typer.scala
@@ -11,6 +11,7 @@ import effekt.source.{ AnyPattern, Def, Expr, IgnorePattern, MatchPattern, Modul
 import effekt.substitutions._
 import effekt.symbols._
 import effekt.symbols.builtins._
+import effekt.symbols.kinds._
 import effekt.util.messages.FatalPhaseError
 import org.bitbucket.inkytonik.kiama.util.Messaging.Messages
 
@@ -426,7 +427,7 @@ class Typer extends Phase[ModuleDecl, ModuleDecl] {
         Context.assignType(field, tpe)
       }
 
-    case d: source.TypeDef   => wellformedType(d.symbol.tpe)
+    case d: source.TypeDef   => wellformed(d.symbol.tpe)
     case d: source.EffectDef => wellformed(d.symbol.effs)
     case _                   => ()
   }
@@ -497,91 +498,6 @@ class Typer extends Phase[ModuleDecl, ModuleDecl] {
     case _ => Context.panic("Cannot extract type")
   }
 
-  // just for checking wellformedness...
-  sealed trait Kind
-  object Kind {
-    case object Type extends Kind
-    case object Effect extends Kind
-    case class Fun(params: List[Kind], res: Kind) extends Kind
-    def TypeFun(params: List[Kind]): Kind =
-      if (params.isEmpty) Kind.Type else Kind.Fun(params, Kind.Type)
-    def EffectFun(params: List[Kind]): Kind =
-      if (params.isEmpty) Kind.Effect else Kind.Fun(params, Kind.Effect)
-  }
-
-  def wellformed(tpe: Type)(implicit C: Context): Unit = tpe match {
-    case t: ValueType      => wellformedType(t)
-    case t: BlockType      => wellformed(t)
-    case t: CapabilityType => wellformedEffect(t.eff)
-  }
-
-  def wellformed(b: BlockType)(implicit C: Context): Unit = b match {
-    case BlockType(tparams, params: Sections, ret) =>
-      params.flatten.foreach { tpe => wellformed(tpe) }
-      wellformed(ret)
-  }
-  def wellformed(tpe: ValueType)(implicit C: Context): Kind = tpe match {
-    case FunType(tpe, region) =>
-      wellformed(tpe); Kind.Type
-    case _: TypeVar => Kind.Type
-    case TypeApp(tpe, args) =>
-      val Kind.Fun(params, res) = wellformed(tpe) match {
-        case t: Kind.Fun => t
-        case _           => Context.abort(s"Expected a type constructor, but got: ${tpe}")
-      }
-      if (args.size != params.size) {
-        Context.abort(s"Wrong type constructor arity. Type constructor ${tpe} expects ${params.size} parameters, but got ${args.size} arguments.")
-      }
-      args foreach { a => wellformedType(a) }
-      res
-
-    case TypeAlias(_, tparams, tpe) =>
-      wellformedType(tpe)
-      Kind.TypeFun(tparams map { p => Kind.Type })
-    case DataType(_, tparams, _)  => Kind.TypeFun(tparams map { p => Kind.Type })
-    case Record(_, tparams, _, _) => Kind.TypeFun(tparams map { p => Kind.Type })
-    case BuiltinType(_, tparams)  => Kind.TypeFun(tparams map { p => Kind.Type })
-  }
-  def wellformed(e: Effect)(implicit C: Context): Kind = e match {
-    case EffectApp(eff, args) =>
-      val Kind.Fun(params, res) = wellformed(eff) match {
-        case t: Kind.Fun => t
-        case _           => Context.abort(s"Expected an effect that takes type parameters, but got: ${eff}")
-      }
-      if (args.size != params.size) {
-        Context.abort(s"Wrong number of type arguments. Effect ${eff} expects ${params.size} parameters, but got ${args.size} arguments.")
-      }
-      args foreach { a => wellformedType(a) }
-      res
-    case EffectAlias(_, tparams, effs) =>
-      wellformed(effs)
-      Kind.EffectFun(tparams map { p => Kind.Type })
-    case UserEffect(_, tparams, _) =>
-      Kind.EffectFun(tparams map { p => Kind.Type })
-    case BuiltinEffect(_, tparams) =>
-      Kind.EffectFun(tparams map { p => Kind.Type })
-  }
-  def wellformed(effs: Effects)(implicit C: Context): Unit = effs.toList foreach { eff =>
-    wellformed(eff) match {
-      case Kind.Effect => ()
-      case _           => Context.abort(s"Expected an effect but got ${eff}")
-    }
-  }
-
-  def wellformedType(tpe: ValueType)(implicit C: Context): Unit = wellformed(tpe) match {
-    case Kind.Type => ()
-    case _         => Context.abort(s"Expected a simple type, but got: ${tpe}")
-  }
-
-  def wellformedEffect(eff: Effect)(implicit C: Context): Unit = wellformed(eff) match {
-    case Kind.Effect => ()
-    case _           => Context.abort(s"Expected a simple effect, but got: ${eff}")
-  }
-
-  def wellformed(e: Effectful)(implicit C: Context): Unit = {
-    wellformedType(e.tpe)
-    wellformed(e.effects)
-  }
 
   /**
    * Returns the binders that will be introduced to check the corresponding body
@@ -873,26 +789,13 @@ class Typer extends Phase[ModuleDecl, ModuleDecl] {
       checkStmt(stmt, Some(tpe))
   }
 
-  private implicit class ValueTypeOps[T <: ValueType](tpe: T) {
-    def wellformed(implicit C: Context): T = {
-      wellformedType(tpe)
-      tpe
-    }
-  }
-  private implicit class EffectOps[T <: Effect](eff: T) {
-    def wellformed(implicit C: Context): T = {
-      wellformedEffect(eff)
-      eff
-    }
-  }
-
   /**
    * Combinators that also store the computed type for a tree in the TypesDB
    */
   def checkAgainst[T <: Tree](t: T, expected: Option[Type])(f: T => Effectful)(implicit C: Context): Effectful =
     Context.at(t) {
       val (got / effs) = f(t)
-      wellformedType(got)
+      wellformed(got)
       wellformed(effs)
       expected foreach { Context.unify(_, got) }
       C.assignType(t, got / effs)
@@ -902,7 +805,7 @@ class Typer extends Phase[ModuleDecl, ModuleDecl] {
   def check[T <: Tree](t: T)(f: T => Effectful)(implicit C: Context): Effectful =
     Context.at(t) {
       val (got / effs) = f(t)
-      wellformedType(got)
+      wellformed(got)
       wellformed(effs)
       Context.assignType(t, got / effs)
       got / effs
